requirements.txt
aenum==3.1.16
ai-api-client-sdk==2.4.6
ai-core-sdk==2.5.7
alembic==1.8.1
amqp==5.1.1
async-timeout==4.0.2
awscli==1.40.3
bcrypt==4.0.1
billiard==4.2.1
blinker==1.9.0
boto3==1.38.15
botocore==1.38.15
cachetools==5.5.2
celery==5.5.2
certifi==2022.9.24
cffi==1.17.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
cryptography==38.0.3
Deprecated==1.2.13
dnspython==2.2.1
docopt==0.6.2
docutils==0.19
durationpy==0.9
email-validator==1.3.0
et-xmlfile==1.1.0
execnet==2.1.1
Flask==2.2.2
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Seeder==1.2.0
Flask-SQLAlchemy==3.0.2
flask-swagger-ui==4.11.1
Flask-WTF==1.0.1
flower==1.2.0
google-auth==2.39.0
greenlet==3.2.2
gunicorn==20.1.0
humanize==4.4.0
idna==3.4
iniconfig==2.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
kombu==5.5.2
kubernetes==32.0.1
Mako==1.2.3
MarkupSafe==2.1.1
metaflow==2.15.9
numpy==2.2.5
oauthlib==3.2.2
openpyxl==3.0.10
packaging==21.3
pandas==2.2.3
pipreqs==0.4.11
pluggy==1.6.0
prometheus-client==0.15.0
prompt-toolkit==3.0.33
py-healthcheck==1.10.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.21
Pygments==2.19.2
pyhumps==3.8.0
PyJWT==2.6.0
PyMySQL==1.0.2
pyparsing==3.0.9
pytest==8.4.1
pytest-xdist==3.8.0
flake8==7.1.1
black==24.8.0
isort==5.13.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
pyxlsb==1.0.10
PyYAML==6.0.2
redis==4.3.4
requests==2.28.1
requests-oauthlib==2.0.0
rsa==4.7.2
s3transfer==0.12.0
sap-ai-core-metaflow==1.1.21
setuptools==80.4.0
six==1.16.0
SQLAlchemy==1.4.43
tornado==6.2
tzdata==2025.2
urllib3==1.26.12
vine==5.1.0
wcwidth==0.2.5
websocket-client==1.8.0
Werkzeug==2.2.2
wrapt==1.14.1
WTForms==3.0.1
yarg==0.1.9
